initial organization
