+++ Source/AwsCommonRuntimeKit/mqtt/Mqtt5Packets.swift
@@ -2,6 +2,7 @@
 ///  SPDX-License-Identifier: Apache-2.0.
 
 import Foundation
+import AwsCHttp
 import AwsCMqtt
 
 /// Mqtt5 User Property
